Update CFTDs.py
BUG FIX: the cloud top heights were being calculated incorrectly. It is now fixed.

In the 2D histograms, beta has been replace for beta_belowDGL with the condition of at least 50% of valid values, for robustness.

In the ridgeplots (regular and  rimed vs non rimed cases) the beta_belowML is now also filtered to have at least 50% of valid values, for robustness.