new alternative clean ML height variable
In compute_qvps_new.py and plot_ppis_qvps_etc.py an alternative "clean" version of the ML heights was created after quality checks, instead of replacing the original detection result. This new variable is used in the atten correction and the KDP ML correction, but both ML height versions are included in the output data.