Update compare_calibration_attenuation_adjacent_radars.py
Added more variables to process.

New option to (de)activate reloading and calculation of matching and value extraction.

Added custom attenuation calculation in the plotting section.

Added dictionary with list of valid dates based on matched results.

Added min and max calculations in the ML analogously to the ML bump calculations.

New block (TEST) trying out how to improve ML detection. This was already implemented for QVPs, but not for PPIs. May be deleted in the future.